added support for little endian
Generally restructured how instruction arguments are configured, and in so doing added support for little endian byte order for values written to the byte code. Also added the capability to specify the start and/or end address to write to the binary image, which is useful when generating ROM images.