modified how address gaps are filled
Created support for filling address gaps (address values no assembled code pertaqins to) with a fill value, which is specifiable on the comand options if something other than the default 0 value is desired. This not mitigates the need for the directive ".zerountil". The binary image generated will fill all bytes inbetween (inclusive) the indicated start and end address.