added support for local labels
Local labels start with a period . and are usable in between any two non-local labels or a non-local label and the end of a file or a .org directive. Local labels are most useful in subroutines. This feature was implemented with the future of building binaries based on multiple files, so a concept of file-scoped label was added too. File scoped labels start with a _