Added instruction variants to the ISA model (#5)
Instruction variants allow a specific operand configuration for an instruction to generate wholly different byte code, such as a different prefix value. When an instruction is configured in the ISA configuration file, an option named variants can be added to the instruction configuration. This key should contain a list of 1 or more instruction configurations. When BespokeASM then is parsing an instruction for a given mnemonic, it will first try to match the operands to the base configuration for that mnemonic. If no match is found, it will then attempt to match each variant configuration in order. BespokeASM will generate the byte code for the instruction based on the first variant (base configuration or specified variants) matched.