Support for Indirect Indexed Register addressing mode (#6)
Added support for the indirect indexed register addressing more, which allows a register that be dereferenced for n memory address to have that memory address offset by another value source, notably other registers. This is really a generalization of the Indirect Register addressing mode which allows an address offset by an immediate value. In fact, an immediate value is an allowed offset for the Indirect Indexed Register. Offsets can be nested indirect addressing modes. For example, `[sp+[sp+4]]` is allowed.

Other changes include:

Added support for the `.cstr` data type, which is a null-terminated string. Also converted `.byte` strings to not be null terminated.
Added support of python-style character escapes in data strings, both `.cstr` and `.byte`
Added a number of coding error checks.